+++ src/bin_client/bash-completion.sh
@@ -49,8 +49,4 @@ _tezos-alphanet_complete()
 complete -F _tezos-client_complete tezos-client
 complete -F _tezos-client_complete tezos-admin-client
 complete -F _tezos-client_complete tezos-baker-alpha
-complete -F _tezos-client_complete tezos-endorser-alpha
 complete -F _tezos-client_complete tezos-accuser-alpha
-complete -F _tezos-alphanet_complete mainnet.sh
-complete -F _tezos-alphanet_complete carthagenet.sh
-complete -F _tezos-alphanet_complete tezos-docker-manager.sh